Move some slow tests from `test_stateful.py` into `nocover/`
